Add Task tab in WorkflowDefinitionEditor
This commit introduces a new "Task" tab in the `WorkflowDefinitionEditor` located under `ActivityProperties`. The Task tab contains a checkbox for setting if the workflow should be run asynchronously. Some refinements have been made to the CommonTab and VersionTab files for handling activity updates, and a conditional render of the Task tab has been added to the `ActivityPropertiesPanel`.