Update AddRemoteBackend function and Elsa.Api.Client version
The parameters of the AddRemoteBackend function in different files have been modified to enhance how the Authentication Handler is set. The changes involve assigning 'typeof(AuthenticatingApiHttpMessageHandler)' directly to the 'AuthenticationHandler'. Also, the version of the Elsa.Api.Client library in Elsa.Studio.Core has been updated from '3.0.0-preview.932' to '3.0.0-preview.942'.