Integrate useEffect in WorkflowDefinitionList component
In this update, useEffect hook has been added to the WorkflowDefinitionList component. It's been added to handle an 'editWorkflowDefinition' function which is attached to the component's current DOM element. From now on, a callback will be triggered when a workflow definition is edited, which is passed as a prop from the parent component.