Update Elsa.Api.Client version and branch reference
Updated the Elsa.Api.Client package version from a preview version to a stable one in the Elsa.Studio.Core project. In addition, adjusted the branch reference in GitHub actions workflow from v3.0.1 to main ensuring the project references the stable Elsa.Api.Client version and that the workflow tracks changes on the main branch, respectively.