Workflow Instance Filter Improvements (#151)
* Add 'enhancement/*' branch to Github workflow

The workflow configuration in packages.yml has been updated. The 'enhancement/*' branch pattern is now included in the action triggers. This allows the workflow to run for any enhancements in the codebase.

* Update incident filtering in WorkflowInstanceList

Changed the incident filter option in the WorkflowInstanceList from a checkbox to a dropdown list. This allows the user to select a clearer "Yes" or "No" option instead of a checkbox. In addition, updated the OnHasIncidentsChanged method to be asynchronous.

* Update label for "UpdatedAt" field in WorkflowInstanceList

The label for the "UpdatedAt" field in the WorkflowInstanceList component was changed from "Last executed" to "Updated". This better represents what the field is actually showing, as it refers to the most recent update to the workflow instance, not its last execution.

* Update time filtering in WorkflowInstanceList

The time filter input in the WorkflowInstanceList module has been updated. It is now handled as an Nullable string instead of a non-nullable one. Additionally, the UI has been improved by replacing the MudTimePicker with a clearer MudTextField with an appropriate format and icon.

* Update AnchorOrigin to BottomCenter in WorkflowInstanceList

This commit changes the AnchorOrigin in the WorkflowInstanceList component from BottomLeft to BottomCenter. This alteration should make the user interface more symmetrical and easier to interact with.

* Remove unnecessary identifiers from WorkflowInstanceList

The WorkflowInstanceList no longer checks against the workflow definition's description, ID, and definition ID. This change simplifies the search functionality and focuses on only the workflows' names, which should facilitate user experience and search efficiency.