Update tool and package versions to 3.3.0
The tool version has been updated in ToolVersion.cs. This change is reflected in the base_version environment variable in the GitHub actions configuration file. Additionally, the version in the GitHub actions file is no longer hardcoded but now references this base_version variable, ensuring consistency and easier future updates.